Revert "Introduce solana auth integration"
