--- dto/flowid.go
@@ -1,6 +1,0 @@
-package dto
-
-type FlowId struct {
-	WalletAddress string
-	FlowId        string `gorm:"primary_key"`
-}
