--- internal/pkg/errorso/errorso.go
@@ -1,8 +1,0 @@
-// Package errorso provides common error declarations
-package errorso
-
-import "errors"
-
-var (
-	ErrRecordNotFound = errors.New("record not found")
-)
