+++ logger/logger.go
@@ -9,13 +9,13 @@ import (
 
 var log *zap.Logger
 
-func init() {
+func init(){
 	var err error
-	config := zap.NewProductionConfig()
-	encoderConfig := zap.NewProductionEncoderConfig()
-	encoderConfig.TimeKey = "timestamp"
-	encoderConfig.EncodeTime = zapcore.ISO8601TimeEncoder
-	config.EncoderConfig = encoderConfig
+	config:=zap.NewProductionConfig()
+	encoderConfig:=zap.NewProductionEncoderConfig()
+	encoderConfig.TimeKey="timestamp"
+	encoderConfig.EncodeTime=zapcore.ISO8601TimeEncoder
+	config.EncoderConfig=encoderConfig
 	fileEncoder := zapcore.NewJSONEncoder(config.EncoderConfig)
 	logFile, _ := os.OpenFile("log.json", os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
 	writer := zapcore.AddSync(logFile)
@@ -28,32 +28,20 @@ func init() {
 	log = zap.New(core, zap.AddCaller(), zap.AddStacktrace(zapcore.ErrorLevel))
 
 	// log,err=zap.NewProduction(zap.AddCallerSkip(1))
-	if err != nil {
+	if err!=nil{
 		panic(err)
 	}
 
 }
 
-func Info(message string, fields ...zap.Field) {
-	log.Info(message, fields...)
+func Info(message string,fields ...zap.Field){
+	log.Info(message,fields...)
 }
 
-func Debug(message string, fields ...zap.Field) {
-	log.Debug(message, fields...)
+func Debug(message string,fields ...zap.Field){
+	log.Debug(message,fields...)
 }
 
-func Error(message string, fields ...zap.Field) {
-	log.Error(message, fields...)
-}
-
-func Errorf(template string, args ...interface{}) {
-	log.Sugar().Errorf(template, args...)
-}
-
-func Fatal(message string, fields ...zap.Field) {
-	log.Fatal(message, fields...)
-}
-
-func Fatalf(template string, args ...interface{}) {
-	log.Sugar().Fatalf(template, args...)
+func Error(message string,fields ...zap.Field){
+	log.Error(message,fields...)
 }
